fix: Solved error when user change financial close.
